third_party/WebKit/LayoutTests/fast/dom/htmlallcollection-call-with-index-caching-bug.html

<!DOCTYPE html>
<html>
<head id="foo">
<meta charset="utf-8">
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>

description("This tests verifies that calling document.all(name, index) doesn't affect subsequent calls to document.all.item(index)");

shouldBe("document.all.item(0)", "document.documentElement");
debug("Calling document.all('foo', 0).");
document.all('foo', 0)
shouldBe("document.all.item(0)", "document.documentElement");

</script>
</body>
</html>
