third_party/blink/web_tests/platform/generic/fast/dom/htmlallcollection-call-with-index-caching-bug-expected.txt

This tests verifies that calling document.all(name, index) doesn't affect subsequent calls to document.all.item(index)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.all.item(0) is document.documentElement
Calling document.all('foo', 0).
PASS document.all.item(0) is document.documentElement
PASS successfullyParsed is true

TEST COMPLETE
